Improved MinCount result
